extract CategoryFilter to own component
